Move methods to populate various particle types into RealSpaceBuilderUtils
