+++ GUI/coregui/Views/RealSpaceWidgets/RealSpaceBuilder.cpp
@@ -30,11 +30,7 @@
 #include "Lattice2DItems.h"
 #include "Units.h"
 #include "ParticleCompositionItem.h"
-#include "IParticle.h"
-#include "IFormFactorDecorator.h"
-#include "Particle.h"
 #include "ParticleCoreShellItem.h"
-#include "ParticleCoreShell.h"
 
 RealSpaceBuilder::RealSpaceBuilder(QWidget* parent)
     : QWidget(parent)
@@ -68,10 +64,10 @@ void RealSpaceBuilder::populate(RealSpaceModel* model,
     else if (item.modelType() == Constants::ParticleType)
         populateParticle(model, item);
 
-    else if (item.modelType() == Constants::ParticleCoreShellType)
+    else if (item.modelType() == Constants::ParticleCompositionType)
         populateParticle(model, item);
 
-    else if (item.modelType() == Constants::ParticleCompositionType)
+    else if (item.modelType() == Constants::ParticleCoreShellType)
         populateParticle(model, item);
 }
 
@@ -168,107 +164,18 @@ void RealSpaceBuilder::populateParticle(RealSpaceModel* model,
     {
         auto particleCompositionItem = dynamic_cast<const ParticleCompositionItem*>(&particleItem);
         auto particleComposition = particleCompositionItem->createParticleComposition();
-        populateParticleComposition(model, particleComposition.get(), origin);
+        RealSpaceBuilderUtils::populateParticleComposition(model, particleComposition.get(), origin);
     }
     else if( particleItem.modelType() == Constants::ParticleCoreShellType)
     {
         auto particleCoreShellItem = dynamic_cast<const ParticleCoreShellItem*>(&particleItem);
         auto particleCoreShell = particleCoreShellItem->createParticleCoreShell();
-        populateParticleCoreShell(model, particleCoreShell.get(), origin);
+        RealSpaceBuilderUtils::populateParticleCoreShell(model, particleCoreShell.get(), origin);
     }
     else if(particleItem.modelType() == Constants::ParticleType)
     {
         auto pItem = dynamic_cast<const ParticleItem*>(&particleItem);
         auto particle = pItem->createParticle();
-        auto particle3D = TransformTo3D::createParticle3D(particleItem);
-
-        RealSpaceBuilderUtils::applyParticleTransformations(particle.get(), particle3D.get(), origin);
-        RealSpaceBuilderUtils::applyParticleColor(particle.get(), particle3D.get());
-
-        if (particle3D)
-            model->add(particle3D.release());
+        RealSpaceBuilderUtils::populateSingleParticle(model, particle.get(), origin);
     }
-}
-
-void RealSpaceBuilder::populateParticleComposition(RealSpaceModel *model,
-                                                   const ParticleComposition *particleComposition,
-                                                   const QVector3D &origin) const
-{
-    SafePointerVector<IParticle> pc_vector = particleComposition->decompose();
-
-    for (IParticle* pc_particle : pc_vector)
-    {
-        if(dynamic_cast<ParticleCoreShell*>(pc_particle))
-        {
-            auto particleCoreShell = dynamic_cast<ParticleCoreShell*>(pc_particle);
-            populateParticleCoreShell(model, particleCoreShell, origin);
-        }
-        else
-        {
-            auto particle = dynamic_cast<Particle*>(pc_particle);
-            populateSingleParticle(model, particle, origin);
-        }
-    }
-}
-
-void RealSpaceBuilder::populateParticleCoreShell(RealSpaceModel *model,
-                                                 const ParticleCoreShell *particleCoreShell,
-                                                 const QVector3D &origin) const
-{
-    const IFormFactor* coreParticleff = particleCoreShell->coreParticle()->createFormFactor();
-    const IFormFactor* shellParticleff = particleCoreShell->shellParticle()->createFormFactor();
-
-    // TRUE as long as coreParticleff is of IFormFactorDecorator (or its derived) type
-    while(dynamic_cast<const IFormFactorDecorator*>(coreParticleff))
-        coreParticleff =
-                dynamic_cast<const IFormFactorDecorator*>(coreParticleff)->getFormFactor();
-
-    // TRUE as long as shellParticleff is of IFormFactorDecorator (or its derived) type
-    while(dynamic_cast<const IFormFactorDecorator*>(shellParticleff))
-        shellParticleff =
-                dynamic_cast<const IFormFactorDecorator*>(shellParticleff)->getFormFactor();
-
-    auto coreParticle3D = TransformTo3D::createParticlefromIFormFactor(coreParticleff);
-    auto shellParticle3D = TransformTo3D::createParticlefromIFormFactor(shellParticleff);
-
-    // core
-    RealSpaceBuilderUtils::applyParticleCoreShellTransformations(particleCoreShell->coreParticle(),
-                                                                 coreParticle3D.get(),
-                                                                 particleCoreShell,
-                                                                 origin);
-    RealSpaceBuilderUtils::applyParticleColor(particleCoreShell->coreParticle(),
-                                              coreParticle3D.get());
-
-    // shell (set an alpha value of 0.5 for transparency)
-    RealSpaceBuilderUtils::applyParticleCoreShellTransformations(particleCoreShell->shellParticle(),
-                                                                 shellParticle3D.get(),
-                                                                 particleCoreShell,
-                                                                 origin);
-    RealSpaceBuilderUtils::applyParticleColor(particleCoreShell->shellParticle(),
-                                              shellParticle3D.get(), 0.5);
-
-    if (coreParticle3D)
-        model->add(coreParticle3D.release());
-
-    if (shellParticle3D)
-        model->addBlend(shellParticle3D.release()); // use addBlend() for transparent object
-}
-
-void RealSpaceBuilder::populateSingleParticle(RealSpaceModel *model, const Particle *particle,
-                                              const QVector3D &origin) const
-{
-    const IFormFactor* ff = particle->createFormFactor();
-
-    // TRUE as long as ff is of IFormFactorDecorator (or its derived) type
-    while(dynamic_cast<const IFormFactorDecorator*>(ff))
-        ff = dynamic_cast<const IFormFactorDecorator*>(ff)->getFormFactor();
-
-    auto particle3D = TransformTo3D::createParticlefromIFormFactor(ff);
-
-    RealSpaceBuilderUtils::applyParticleTransformations(particle, particle3D.get(),
-                                                        origin);
-    RealSpaceBuilderUtils::applyParticleColor(particle, particle3D.get());
-
-    if (particle3D)
-        model->add(particle3D.release());
 }
