+++ GUI/coregui/Views/RealSpaceWidgets/RealSpaceBuilder.h
@@ -24,9 +24,6 @@ class SessionItem;
 class RealSpaceModel;
 class Shape3D;
 class SceneGeometry;
-class ParticleCoreShell;
-class ParticleComposition;
-class Particle;
 
 class BA_CORE_API_ RealSpaceBuilder : public QWidget
 {
@@ -59,18 +56,6 @@ public:
 
     void populateParticle(RealSpaceModel* model, const SessionItem& particleItem,
                           const QVector3D& origin = QVector3D()) const;
-
-    void populateParticleComposition(RealSpaceModel *model,
-                                     const ParticleComposition* particleComposition,
-                                     const QVector3D &origin) const;
-
-    void populateParticleCoreShell(RealSpaceModel *model,
-                                   const ParticleCoreShell* particleCoreShell,
-                                   const QVector3D &origin) const;
-
-    void populateSingleParticle(RealSpaceModel *model, const Particle* particle,
-                                const QVector3D &origin) const;
-
 };
 
 #endif // REALSPACEBUILDER_H
